exercise 02
Exercise 02: like the exercise 01, but adding multiplication and division. Must be used a function with 3 values: number 1, number 2 and the arithmetic. List the possibles arithmetics (a letter to each one). You can not use else. The function can have any number of returns needed. Show a error message if the arithmetic doesn't exist.